+++ app/flows/prozesskostenhilfeFormular/index.ts
@@ -1,5 +1,8 @@
 import _ from "lodash";
 import type { BasicTypes } from "~/flows/contexts";
+import type { ProzesskostenhilfeFinanzielleAngabenEinkuenfteContext } from "~/flows/prozesskostenhilfeFormular/finanzielleAngaben/einkuenfte/context";
+import { einkuenfteDone } from "~/flows/prozesskostenhilfeFormular/finanzielleAngaben/einkuenfte/doneFunctions";
+import { finanzielleAngabeEinkuenfteGuards } from "~/flows/prozesskostenhilfeFormular/finanzielleAngaben/einkuenfte/guards";
 import abgabeFlow from "./abgabe/flow.json";
 import { prozesskostenhilfeAbgabeGuards } from "./abgabe/guards";
 import type { ProzesskostenhilfeFinanzielleAngabenContext } from "./finanzielleAngaben/context";
@@ -7,10 +10,10 @@ import {
   andereUnterhaltszahlungenDone,
   eigentumDone,
   eigentumZusammenfassungDone,
-  einkuenfteDone,
   kinderDone,
   partnerDone,
 } from "./finanzielleAngaben/doneFunctions";
+import einkuenfteFlow from "./finanzielleAngaben/einkuenfte/flow.json";
 import finanzielleAngabenFlow from "./finanzielleAngaben/flow.json";
 import { finanzielleAngabeGuards } from "./finanzielleAngaben/guards";
 import prozesskostenhilfeFormularFlow from "./flow.json";
@@ -81,7 +84,9 @@ export const prozesskostenhilfeFormular = {
       start: { meta: { done: () => true } },
       "finanzielle-angaben": _.merge(finanzielleAngabenFlow, {
         states: {
-          einkuenfte: { meta: { done: einkuenfteDone } },
+          einkuenfte: _.merge(einkuenfteFlow, {
+            meta: { done: einkuenfteDone },
+          }),
           partner: { meta: { done: partnerDone } },
           kinder: { meta: { done: kinderDone } },
           "andere-unterhaltszahlungen": {
@@ -100,6 +105,7 @@ export const prozesskostenhilfeFormular = {
   }),
   guards: {
     ...finanzielleAngabeGuards,
+    ...finanzielleAngabeEinkuenfteGuards,
     ...prozesskostenhilfeAbgabeGuards,
   },
   stringReplacements: (context: ProzesskostenhilfeFormularContext) => ({
@@ -112,4 +118,6 @@ export const prozesskostenhilfeFormular = {
 } as const;
 
 export type ProzesskostenhilfeFormularContext =
-  ProzesskostenhilfeFinanzielleAngabenContext & AbgabeContext;
+  ProzesskostenhilfeFinanzielleAngabenContext &
+    ProzesskostenhilfeFinanzielleAngabenEinkuenfteContext &
+    AbgabeContext;
